CRM beta

+++ BRT/src/main/java/rom/brt/dto/Subscriber.java
@@ -3,10 +3,10 @@ package rom.brt.dto;
 import java.time.LocalDate;
 
 public record Subscriber(
-        Integer id,
+        Long id,
         String msisdn,
         boolean isServiced,
-        Integer tariffId,
+        Long tariffId,
         Integer minutes,
         LocalDate paymentDay
 ) {
@@ -22,7 +22,7 @@ public record Subscriber(
         }
     }
 
-    public static Subscriber fromServicedUser(int id, String msisdn, int tariffId, int minutes, LocalDate paymentDay) {
+    public static Subscriber fromServicedUser(long id, String msisdn, long tariffId, int minutes, LocalDate paymentDay) {
         return new Subscriber(id, msisdn, true, tariffId, minutes, paymentDay);
     }
